+++ loop-templates/templates/src/whispers/index.ts
@@ -15,6 +15,8 @@ import networkWhisperTest from './NetworkWhisper.test.ts.squirrelly';
 import networkWhisper from './NetworkWhisper.ts.squirrelly';
 import processWhisperTest from './ProcessWhisper.test.ts.squirrelly';
 import processWhisper from './ProcessWhisper.ts.squirrelly';
+import systemWhisperTest from './SystemWhisper.test.ts.squirrelly';
+import systemWhisper from './SystemWhisper.ts.squirrelly';
 import uiWhisperTest from './UiWhisper.test.ts.squirrelly';
 import uiWhisper from './UiWhisper.ts.squirrelly';
 import userWhisperTest from './UserWhisper.test.ts.squirrelly';
@@ -67,6 +69,11 @@ const fileMap: FileMap = {
     aptitude: 'process',
   },
   processWhisper: { fileName: 'ProcessWhisper.ts', aptitude: 'process' },
+  systemWhisperTest: {
+    fileName: 'SystemWhisper.test.ts',
+    aptitude: 'system',
+  },
+  systemWhisper: { fileName: 'SystemWhisper.ts', aptitude: 'system'},
   uiWhisperTest: { fileName: 'UiWhisper.test.ts', aptitude: 'ui' },
   uiWhisper: { fileName: 'UiWhisper.ts', aptitude: 'ui' },
   userWhisperTest: { fileName: 'UserWhisper.test.ts', aptitude: 'user' },
@@ -95,6 +102,8 @@ export default {
   networkWhisper,
   processWhisperTest,
   processWhisper,
+  systemWhisperTest,
+  systemWhisper,
   uiWhisperTest,
   uiWhisper,
   userWhisperTest,