+++ loop-templates/templates/types.ts
@@ -16,6 +16,7 @@ export type TemplateFile = {
     | 'keyboard'
     | 'network'
     | 'process'
+    | 'system'
     | 'ui'
     | 'user'
     | 'vault'